python/requirements.txt
fastapi[standard]
uvicorn
pydantic
google.generativeai
python-dotenv
pandas
scikit-learn
numpy
fpdf2